backend/requirements.txt
# FastAPI and Server
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.12

# Vector Database and Embeddings
# Try newer chromadb version with better wheel support
chromadb==0.5.20
sentence-transformers==2.7.0

# Data Processing
beautifulsoup4==4.12.3
requests==2.32.3
lxml==5.3.0
tiktoken==0.8.0

# Environment and Configuration
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.7.1

# LLM Providers
groq==0.11.0
google-generativeai==0.8.3

# Utilities
python-dateutil==2.9.0.post0

# NumPy - pin to version with wheels for Python 3.14
numpy==2.1.1

# AssemblyAI SDK (for reliable STT)
assemblyai>=0.28.0



